requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alphashape==1.3.1
anyio==3.6.2
appdirs==1.4.4
apptools==5.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
async-generator==1.10
async-timeout==5.0.1
attrs==25.3.0
autobahn==21.2.1
Automat==22.10.0
Babel==2.11.0
backcall==0.2.0
bleach==4.1.0
blinker==1.9.0
cached-property==1.5.2
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
click==8.1.8
click-log==0.4.0
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.8
constantly==15.1.0
contextvars==2.4
contourpy==1.3.0
cryptography==40.0.2
dash==3.0.3
dataclasses==0.8
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.4
envisage==6.1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.57.0
frozenlist==1.5.0
h5py==3.1.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
immutables==0.19
importlib-resources==5.4.0
importlib_metadata==8.6.1
incremental==22.10.0
ipykernel==5.5.6
ipython==7.16.0
ipython-genutils==0.2.0
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==0.17.0
json5==0.9.16
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==7.1.2
jupyter-packaging==0.10.6
jupyter-server==1.13.1
jupyter_core==5.7.2
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab_widgets==3.0.14
lazy_loader==0.4
llvmlite==0.36.0
MarkupSafe==3.0.2
matlabengineforpython===R2020b
matplotlib==3.3.4
matplotlib-inline==0.1.7
mayavi==4.8.0
meshio==4.4.6
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.1.0
multidict==6.4.3
narwhals==1.35.0
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
notebook==6.4.10
numba==0.53.1
numpy==1.19.5
numpy-stl==3.2.0
olefile==0.46
open3d==0.15.1
openpyxl==3.0.9
packaging==24.2
pandas==1.1.5
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pickleshare==0.7.5
Pillow==8.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.7
plotly==6.0.1
pooch==1.8.2
prometheus-client==0.17.1
prompt_toolkit==3.0.50
propcache==0.3.1
pure_eval==0.2.3
pycparser==2.21
pyface==7.4.4
Pygments==2.19.1
pynrrd==0.4.3
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
pyvista==0.32.1
PyWavelets==1.1.1
pywin32==305
pywinpty==1.1.6
pyzmq==25.1.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
rtree==1.4.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
scooby==0.5.11
seaborn==0.4.0
Send2Trash==1.8.3
shapely==2.0.7
SimpleITK==2.1.0
six==1.17.0
slicer==0.0.8
sniffio==1.2.0
stack-data==0.6.3
statsmodels==0.12.2
terminado==0.12.1
testpath==0.6.0
threadpoolctl==2.2.0
tifffile==2024.8.30
tomlkit==0.11.6
traitlets==5.14.3
traits==6.4.1
traitsui==7.4.3
trimesh==4.6.6
Twisted==22.4.0
twisted-iocpsupport==1.0.4
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2025.2
urllib3==2.4.0
vedo==2025.5.3
vtk==9.0.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wincertstore==0.2
wslink==2.3.3
xlrd==2.0.1
yarl==1.19.0
zipp==3.21.0
zope.interface==5.5.2
